simplify type proxy
